requirements/development.txt
Django==5.0.6
black==24.10.0
django-environ==0.11.2
django-filter==23.1
django-redis==5.2.0
djangorestframework==3.15.2
factory_boy==3.3.1
gevent==24.11.1
gunicorn==23.0.0
ipdb==0.13.13
isort==5.13.2
psycopg2-binary==2.9.10